sysklogd: fix update-rc.d handling
The sysklogd recipe had a cut-n-paste version of the
update-rc.d.bbclass code which didn't work, but this was hidden
because all images contain the busybox version which does.  Building a
busybox-free image unmasked the issue and syslogd wouldn't start on
first boot.

The comments seem to be wrong/stale.  AFAICT update-rc.d and
update-alternatives work fine with each other, though there is an
ordering constraint (alternatives must be specified last, so it
"wraps" update-rc.d).  This version builds and works both with and
without busybox.

Signed-off-by: Andy Ross <andy.ross@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>