bitbake.conf: Change build output message to list BUILD_SYS, TARGET_SYS and NATIVELSBSTRING
The build summary is meant to reflect key configuration variables. Information
about the build system we're running on is important but currently missing
from the information displayed.

Printing TARGET_SYS removes the need to print TARGET_OS and TARGET_ARCH
and we add BUILD_SYS and NATIVELSBSTRING to show information about the
build system.

[YOCTO #3456]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>