populate_sdk_deb.bbclass: put back INSTALL_PACKAGES_NORMAL_DEB
This commit (0b0674ebcd8a51783e4bb38e3ead3e419dbba376) was also changing
the populate_sdk_deb.bbclass but the deb backend does not support
multilib just yet. So, when compiling the standalone toolchain tarball,
we ended up with an empty tarball if the deb package manager was chosen.

Reverting the deb part until the multilib support is added to deb.

[YOCTO #3532]

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>