x11vnc: bind vnc server to port 5900
At the moment, if one clicks on X11VNC Server icon, a new vnc server is
started on an automatically assigned port (starting with 5900). The
problem is that the user can click multiple times on the icon and a new
vnc server is started on another port. So, we'll end up with multiple
vnc servers, wasting memory.

This patch will restrict the x11vnc server port to 5900 (default VNC
port) and another process will not be started (because the port is
already used), unless the user chooses to start the server manually,
from the command line, in which case the port can be changed as the
user wishes.

[YOCTO #3369]

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>