util-linux: moved 'reset' into own subpackage
The 'reset' tool from util-linux is a shell script calling 'tput' which is
provided by 'ncurses'.  To avoid a 'util-linux' -> 'ncurses' dependency,
'reset' is put into an own subpackage which can be installed on demand.

Because most (all) systems are using 'busybux' which provides an
alternative 'reset' implementation, this should not cause too much
regressions.

Signed-off-by: Enrico Scholz <enrico.scholz@informatik.tu-chemnitz.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>