lttng-2.0: components updated to version 2.1.0
lttng2-ust, lttng-modules and lttng-tools were updated to version
2.1.0. Since these should be changed together, the changes are
grouped into a single patch.

The text for the lttng-tools LICENSE file has these lines changed
(according to the license text, these shouldn't impact the actual
licensing terms of the package):

=================================================================
-) include/lttng/*
-) src/lib/lttng-ctl/*

We have some tests LGPL but should not impact anything even if we change them
in the future to GPL.
=================================================================

The text for the lttng-modules LICENSE files contains a minor change in the
MIT section (a reference to the mit-license.txt file):

=================================================================
These files are licensed under an MIT-style license. See mit-license.txt
for details.
=================================================================

Signed-off-by: Bogdan Marinescu <bogdan.a.marinescu@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>