gtk+3: add 3.4.4
Mostly taken from meta-gnome, with some changes.

Instead of depending on gtk+-native, depend on just gdk-pixbuf-native and build
the tool required for the host directly.  This saves building the entire GTK+
stack natively just for one tool.

Append a version suffix and use update-alternatives on gtk-update-icon-cache as
with GTK+ 2.

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>