lttng2-ust: rename to lttng-ust
recipes-kernel/lttng/lttng-ust was removed by a previous patch, which
also removed the naming conflict that caused the lttng-2.0 version of
lttng-ust to be named lttng2-ust.  lttng2-ust was the only lttng-2.0
recipe to be named in this way and since that was entirely due to the
conflict, we can go back to the consistent naming for it.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>