dropbear: Deal with truncated host keys by removing them
Dropbear does not start when the host key is empty and it is possible
that a device is switched off before the host key is generated. This
is possible because the dropbearkey code doesn't create a temporary
file first. Detect truncated keys and then remove them which will lead
to the re-generation. This way the dropbear process will always start.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>