libpcap: Turn canusb support into PACKAGECONFIG
libpcap notices libusb in its configure if its sysrooted
and ignores it if its not. But this causes problems since
there is no explicit dependency on libusb. Rebuild from shared
state sometimes fails when libusb has not been yet populated
in sysroot and libpcap has. And do_configure of consumers of libpcap e.g. tcpdump
are probing for features on libpcap. All the tests fail since
none of link steps succeed and it sees as if libpcap does
not have required features and configure of tcpdump bails out.

turn canusb into PACKAGECONFIG and let it be disabled
by default.

Backport additional patches needed for disable-canusb to work

Signed-off-by: Khem Raj <raj.khem@gmail.com>