core-image: Add hwcodecs IMAGE_FEATURE
Currently hardware codecs are being injected in rather ugly ways and end up
either in no or all images. This adds a dedicated IMAGE_FEATURE for them based
on the MACHINE_HWCODECS variable. We may need to refine this in due course
but this patch at least illustrates the concept.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>