systemd: remove usbutils dependancy
V2:

Dependancy was removed in version 196 according to following systemd commit.

http://cgit.freedesktop.org/systemd/systemd/commit/configure.ac?id=796b06c21b62d13c9021e2fbd9c58a5c6edb2764

V1:

No where in the systemd readme does it specify that usbutils
is a dependancy of systemd. I can only guess that it was added
due to an issue elsewhere in the chain or it was brought in
when udev was merged and the dependancies never checked.

I have build tested this with no issues.

Signed-off-by: Jack Mitchell <jack.mitchell@dbbroadcast.co.uk>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>